resume-relevance-system/requirements.txt
streamlit==1.28.0
google-generativeai==0.3.2
pandas==2.1.4
plotly==5.18.0
PyMuPDF==1.23.8
python-docx==1.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
